style: Replace nav bar template style with tailwind
