+++ src/components/Atoms/Button/index.tsx
@@ -1,2 +1,2 @@
-export { Button } from './Button';
-export type { ButtonProps, buttonVariants } from './Button.type';
+export { Button, buttonVariants } from './Button';
+export type { ButtonProps } from './Button.type';
